requirements.txt
yfinance>=0.2.0
transformers>=4.35.0,<4.45.0
torch>=2.0.0,<2.5.0
torchvision>=0.15.0,<0.20.0
requests>=2.31.0
asyncpraw>=7.7.0
schedule>=1.2.0
httpx>=0.24.0
python-dotenv>=1.0.0
sambanova>=1.0.0
